Panic on concurrent reads and writes
This was initially not included because the race detector would
complain about concurrent reads and writes of of m.flags: Map.Iter
does an atomic write of m.flags, while Map.Get and Iter.Next would do
non-atomic reads.

But, this is not actually unsafe because we don't care that the read
of m.flags is not atomic. It may get a stale value, but that is fine
for the purpose of a cheap check of unsafe concurrent access of the
map.