Use math/rand/v2 in go1.22
Replace unsafe linking to runtime.fastrand64 with math/rand/v2.Uint64.
The global rand instance in math/rand/v2 no longer requires grabbing a
mutex in the fast path, so there's no longer a chance of contention
between other users of the global rand instance.

Performance is still worse, ~2.5ns becomes ~9.5ns, but this is not
likely to be a bottleneck as there is only one call to rand64() per
Iter() call. This feels like a worthwhile tradeoff to remove an unsafe
import and a go:linkname.